python-api/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
python-jose[cryptography]
argon2-cffi
pydantic
pydantic-settings
email-validator
alembic
psycopg2-binary
